Module bonsaidb_server::api

Types for defining API handlers.

Structs

HandlerSession

A session for a Handler, providing ways to access the server and connected client.

Enums

HandlerError

An error that can occur inside of a Backend function.

Traits

Handler

A trait that can dispatch requests for a [Api].

Type Definitions

HandlerResult

The return type from a Handler’s handle() function.