Trait bonsaidb_server::Backend

pub trait Backend: Debug + Send + Sync + Sized + 'static {
    type CustomApi: CustomApi;
    type CustomApiDispatcher: Dispatcher<<Self::CustomApi as CustomApi>::Request, Result = Result<<Self::CustomApi as CustomApi>::Response, BackendError<<Self::CustomApi as CustomApi>::Error>>> + Debug;
    fn dispatcher_for(
        server: &CustomServer<Self>, 
        client: &ConnectedClient<Self>
    ) -> Self::CustomApiDispatcher;

    fn initialize<'life0, 'async_trait>(
        server: &'life0 CustomServer<Self>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn client_connected<'life0, 'life1, 'async_trait>(
        client: &'life0 ConnectedClient<Self>, 
        server: &'life1 CustomServer<Self>
    ) -> Pin<Box<dyn Future<Output = ConnectionHandling> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn client_disconnected<'life0, 'async_trait>(
        client: ConnectedClient<Self>, 
        server: &'life0 CustomServer<Self>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn client_authenticated<'life0, 'async_trait>(
        client: ConnectedClient<Self>, 
        server: &'life0 CustomServer<Self>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }
}

Tailors the behavior of a server to your needs.

Associated Types

type CustomApi: CustomApi

The custom API definition. If you do not wish to have an API, () may be provided.

type CustomApiDispatcher: Dispatcher<<Self::CustomApi as CustomApi>::Request, Result = Result<<Self::CustomApi as CustomApi>::Response, BackendError<<Self::CustomApi as CustomApi>::Error>>> + Debug

The type that implements the Dispatcher trait.

Required methods

fn dispatcher_for(
server: &CustomServer<Self>,
client: &ConnectedClient<Self>
) -> Self::CustomApiDispatcher

Returns a dispatcher to handle custom api requests. The server and client parameters are provided to allow the dispatcher to have access to them when handling the individual actions.

Provided methods

fn initialize<'life0, 'async_trait>(
server: &'life0 CustomServer<Self>
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Invoked once upon the server starting up.

fn client_connected<'life0, 'life1, 'async_trait>(
client: &'life0 ConnectedClient<Self>,
server: &'life1 CustomServer<Self>
) -> Pin<Box<dyn Future<Output = ConnectionHandling> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

A client disconnected from the server. This is invoked before authentication has been performed.

fn client_disconnected<'life0, 'async_trait>(
client: ConnectedClient<Self>,
server: &'life0 CustomServer<Self>
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

A client disconnected from the server.

fn client_authenticated<'life0, 'async_trait>(
client: ConnectedClient<Self>,
server: &'life0 CustomServer<Self>
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

A client successfully authenticated.